+++ projects/lib/src/moveevaluation.h
@@ -61,7 +61,7 @@ class LIB_EXPORT MoveEvaluation
 		 * How many nodes were searched?
 		 * \note For human players this is always 0.
 		 */
-		int nodeCount() const;
+		quint64 nodeCount() const;
 
 		/*!
 		 * The principal variation.
@@ -88,7 +88,7 @@ class LIB_EXPORT MoveEvaluation
 		void setTime(int time);
 
 		/*! Sets the node count to \a nodeCount. */
-		void setNodeCount(int nodeCount);
+		void setNodeCount(quint64 nodeCount);
 
 		/*! Sets the principal variation to \a pv. */
 		void setPv(const QString& pv);
@@ -98,7 +98,7 @@ class LIB_EXPORT MoveEvaluation
 		int m_depth;
 		int m_score;
 		int m_time;
-		int m_nodeCount;
+		quint64 m_nodeCount;
 		QString m_pv;
 };
